requirements.txt
requests
python-telegram-bot==13.15
APScheduler==3.6.3